Modifications to allow for MSVC-2008 64bits compilation on windows.
git-svn-id: svn://svn.code.sf.net/p/qtvlm/code/trunk@1209 2387b160-4621-4e8a-be36-40ea10d72da2

+++ src/GshhsPolyReader.cpp
@@ -339,7 +339,7 @@ void GshhsPolyReader::readPolygonFileHeader(FILE *polyfile, PolygonFileHeader *h
     int FReadResult = 0;
 
     fseek(polyfile, 0 , SEEK_SET);
-    FReadResult = fread(header, sizeof(PolygonFileHeader), 1, polyfile);
+    FReadResult = (int)fread(header, sizeof(PolygonFileHeader), 1, polyfile);
 }
 
 //-------------------------------------------------------------------------
